Deferred deletion check: support RHEL7.4+ kernels
These kernels have the feature for deferred deletion built-in
but disabled by default. A sysctl knob is available to turn it on.

This patch checks if the knob is available but not enabled -- in
this case case it enables it for the duration of the test and gives
a user a recommendation on how to enable it permanently.

	BEFORE:
```
[root@kir-ce73-gd docker-devicemapper-setup]#
./deferred_deletion_check/standalone_deferred_deletion_check.sh
Kernel version: 3.10.0-693.2.2.el7.x86_64
Deferred deletion is not supported
```
	AFTER:
```
[root@kir-ce73-gd docker-devicemapper-setup]#
./deferred_deletion_check/standalone_deferred_deletion_check.sh
Kernel version: 3.10.0-693.2.2.el7.x86_64

WARNING: it seems you are using RHEL/CentOS 7.4+ kernel but the
fs.may_detach_mounts sysfs setting is disabled (set to 0).

Setting fs.may_detach_mounts = 1 for the duration of the test.

To enable this permanently, run the following:

	echo "fs.may_detach_mounts=1" | sudo tee -a
/etc/sysctl.d/90-docker.conf
	sudo sysctl -f /etc/sysctl.d/90-docker.conf

Deferred deletion is supported
```

[v2: drop use of trap as it's used elsewhere]

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>